BitParser: use intptr_t where appropriate
intptr_t is an integer type capable of holding a pointer

error: cast from ‘unsigned char*’ to ‘unsigned int’ loses precision [-fpermissive]

+++ BitParser.cpp
@@ -118,7 +118,7 @@ int BitParser::nextChangingElement()
 	}
 	// parse untill we are aligned
 	for (c++;
-	     (c < sentinel) && (((unsigned int) c % sizeof(word)) != 0);
+	     (c < sentinel) && (((intptr_t) c % sizeof(word)) != 0);
 	     c++) {
 		if (*c != run[lastBit])
 			break;
